+++ avm/res/machine-learning-services/workspace/CHANGELOG.md
@@ -2,11 +2,14 @@
 
 The latest version of the changelog can be found [here](https://github.com/Azure/bicep-registry-modules/blob/main/avm/res/machine-learning-services/workspace/CHANGELOG.md).
 
-## 0.12.2
+## 0.13.0
 
 ### Changes
 
-- Regenerate main.json with latest bicep version 0.36.177.2456
+- Add support for datastores
+- Updated ReadMe with AzAdvertizer reference
+- Updated LockType to 'avm-common-types version' `0.6.0`, enabling custom notes for locks.
+- Updated `privateEndpoints` parameter type to 'avm-common-types' `0.6.1`, adding a type to its `tags` property
 
 ### Breaking Changes